tsrelay: break up routes into their own files
