+++ go.mod
@@ -3,6 +3,7 @@ module github.com/tailscale-dev/vscode-tailscale
 go 1.20
 
 require (
+	github.com/go-chi/chi/v5 v5.0.10
 	github.com/gorilla/websocket v1.5.0
 	github.com/mitchellh/go-ps v1.0.0
 	golang.org/x/exp v0.0.0-20230626212559-97b1e661b5df
