+++ tsrelay/handler/auth_middleware.go
@@ -1,0 +1,32 @@
+package handler
+
+import "net/http"
+
+func (h *handler) authMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		user, _, ok := r.BasicAuth()
+
+		// TODO: consider locking down to vscode-webviews://* URLs by checking
+		// r.Header.Get("Origin") only in production builds.
+		w.Header().Set("Access-Control-Allow-Origin", "*")
+		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, DELETE")
+		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
+
+		if r.Method == http.MethodOptions {
+			// Handle preflight request
+			w.WriteHeader(http.StatusNoContent)
+			return
+		}
+
+		if !ok {
+			w.Header().Set("WWW-Authenticate", `Basic realm="restricted", charset="UTF-8"`)
+		}
+
+		if user != h.nonce {
+			// TODO: return JSON for all errors
+			http.Error(w, "unauthorized", http.StatusUnauthorized)
+			return
+		}
+		next.ServeHTTP(w, r)
+	})
+}
